p4: one of the factors is a multiple of 11
This patch steps by 11 in the x loop instead of stepping by 1, taking
advantage of the fact that one of the factors of the palindrome must be
a multiple of 11.

Proof:
Take as given that the largest palindrome that can be written as the
product of two m-digit numbers will have 2m digits. It seems reasonable,
but it might not exist for every m. For this problem, it's given when
m = 3.

Claim 1:
11 divides 10^{2m-1} + 1 for any positive integer m.

Using a telescoping sum, note that
10^{2m-1} + 1 = 10^{2m-1}
                + 10^{2m-1-1} - 10^{2m-1-1}
                + 10^{2m-1-2} - 10^{2m-1-2}
                  ...
                + 10 - 10
                + 1
              = (10 + 1) (\sum_{j=0}^{2m-2} (-1)^{j} * 10^{j})

Claim 2:
11 divides 10^{2m-1-k} + 10^{k} for any positive integer m and any k
with 0 <= k < m.

Factoring out 10^{k},
10^{2m-1-k} + 10^{k} = 10^{k} * (10^{2m-1-2k} + 1)
                     = 10^{k} * (10^{2(m-k)-1} + 1)
The factor on the right is in the form above. So 11 divides
10^{2m-1-k} + 10^{k} for any integers m > 0 and 0 <= k < m.

Let N = x * y be a palindrome where x and y have m digits and N has 2m
digits, so the last exponent of N is 2m-1. Write it as
N = a_{2m-1} * 10^{2m-1} + a_{2m-2} * 10^{2m-2} + ... + a_{1} * 10 + a_{0}
  = \sum_{k=0}^{2m-1} a_{k} * 10^{k}

Since N is a palindrome, a_{k} == a_{2m-1-k} for all 0 <= k < m. So
N = a_{0} * 10^{2m-1-0} + a_{1} * 10^{2m-1-1} + ... + a_{1} * 10 + a_{0}
  = \sum_{k=0}^{m-1} a_{k} * (10^{2m-1-k} + 10^{k})

The factor on the right in the summand has the form in the claims above,
so 11 | N. Since 11 is prime, one of the factors of N is also divisible
by 11, call it x. Use this fact in the loop over values to check.