util: improve logic in isprime()
There are three cases that were previously handled badly when
determining if the input integer n is prime.

1. If n is less than the last prime in the current list, primality
is determined by (non)membership in the list.

2. If the current largest prime is less than n but greater than the
square root of n, primality can be determined by looking for prime
divisors of n in the current list.

3. If the current largest prime is less than the square root of n,
then the list is extended one prime at a time until either a divisor
of n is found or the square root of n is reached.