p97: implement recursive formula with non-recursive loop
Compute base**exponent by using the base 2 representation of
exponent.

To get get the necessary sequence, work backward from base**exponent
and note which of two steps is used:

Let x = exponent and let f be an empty bitstring that will hold a
flag for each step.

While x > 0:
If x is even, set x = x / 2 and prepend 0 to f. This is equivalent to
taking the square root of base**x.
If x is odd, set x = floor(x / 2) and prepend 1 to f. This is
equivalent to diving base**x by base and then taking the square root.

Then to create base**exponent, start with 1 and follow the steps
described by f reading left to right (opposite direction than written):

Let p = 1. Looping over the bits in f:
If 0, set p = p * p
If 1, set p = p * p * base

Note that the bitstring f equals the binary representation of exponent:
Squaring base**exponent is the same as doubling exponent, which is done
by bitshifting exponent left by 1, filling with 0.
Squaring base**exponent and multiplying by base is the same as doubling
exponent then adding 1, or bitshifting exponent left by 1 and filling
with 1.